Revert previous change to keep in sync with Tile codebase. The minor warnings from GCC fixed in the previous revision were false positives due to improper use of "-Wconversion".
Still compiles cleanly with "gcc -Wall -Werror" (gcc 3.3.5).

+++ generic/ttk/ttkTheme.c
@@ -9,7 +9,7 @@
  * See the file "license.terms" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
  *
- * $Id: ttkTheme.c,v 1.5 2007/01/11 15:35:40 dkf Exp $
+ * $Id: ttkTheme.c,v 1.6 2007/01/11 19:59:26 jenglish Exp $
  */
 
 #include <stdlib.h>
@@ -1015,7 +1015,7 @@ int InitializeElementRecord(
 	    *dest = styleDefault ? styleDefault : elementDefault;
 	}
 
-	if (!AllocateResource(cache, tkwin, dest, (int) elementOption->type)) {
+	if (!AllocateResource(cache, tkwin, dest, elementOption->type)) {
 	    return 0;
 	}
     }
